Add fixes to allow files over 2GB to be uploaded
Change the UploadHandler.php file to fix two issues:
Change the value used to fix the overflow from PHP_INT_MAX to the size of a
32 bit integer.  Due to this fix, we also only check the value of the size
via '==' instead of the type and value of size '==='.

This change will likely require updates to the php.ini of the webserver
where the OTJ is running. Update the values of the PHP variables listed below
with the recommended value:

upload_max_filesize: 4096M
post_max_size: 4096M
max_input_time: 10800
max_execution_time: 10800