Add new field to search string
To allow for substring searches, Add a new field to search called
ngram_search which is a copy of the name and description fields.  This
field should use the EdgeNGram tokenizer instead of the whitespace one.

Add this ngram_search field to the Journal code which builds the query
from the search box.