Replace for loop with database query
Replace the "For loop" used to acquire all of the most recent
revision id values and replace it with a query to the itemrevision
table.

The query should pull out the highest revision number for each item_id
value and sort them so that the highest revision number is first.