Fix Searching to match new query
Move the itemrevision table query away from the bottom so that all search
capability is restored.  Change the format of the itemIDs variable to be
consistent with the return of the sql array.

When pushing this change, the cache file should be removed to repoulate it
with the correct format.