Initial commit of OTJ testing
The test is a simple Selenium file to test the main page of the OTJ.
It just walks through the search bar, each of the category filters, and
the issues.  It clicks on each entry and makes sure that the number of
resource links after it is clicked is different than before.