Add CSS to adjust for search bar shifting
Add the css to make the "Content" div of the page have a  float style
This will ensure that it will adjust when the browser window is made
small enough that the search bar drops to a second line.

This is most visible on a submission page of the current functionality.