Add cookies to maintain searches across navigating
Add a cookie called "pastSearch" which will keep a record of the
item ids that were displayed most recently.  This will allow
a user to select an object and go back to the the same search
results as prior to the selection.  Clicking the "clear" button of
the search bar or clicking on the "Home" navigations
will remove the cookie.

Add a second cookie for the search parameters that was used which
will allow the navigation page to display which paramters were used
when navigating back to the page.

Each cookie will last for 30 minutes or until cleared.