Remove parsing of "Public_Report" directory
Add a check which prevents the parsing of the files within the
"Public_Report" directory.  This report is a series of HTML files
that talk about the code which was redacted from a CIF submission.

The licensing and copyright information in these report files makes
it difficult to find the messages from the source of the submission.

For example: http://code.osehra.org/journal/journal/survey?id=885