Add pgSQL connection for handle update
Add a set of commands to update the PostgreSQL database that is used by
the OSEHRA Handle server to determine which URL to open when accessing a
handle link.

Ensure that the command checks for an existing value first and that the error
condition is propagated to the page for display.