Prevent "Private" folder submission from showing
Prevent the public files that are uploaded to the OTJ community folder
named "Private" from being shown on the main page of the OTJ.  This
showed when evidence files from the reviews were found as submissions on
the index page.

Ensure that each item that is displayed has a description.